Introducing the enclave binary
For now as a first step, this runs as a normal process, spawned by the service. This is obviously insecure, but is useful for testing, and is called a "mock" ecnclave.

Vote mixing is now sent off to the enclave process via a UNIX socket. Hopefully, integrating with AWS Nitro Enclaves should just be implementing the VSOCK connection. Otherwise, the enclave binary should remain the same.