Introducing Glove proof
The enclave now creates a random Ed25519 signing key on startup, which it uses to sign the output of its vote mix. The signing key is added to the enclave's attestation doc, and the necessary data types created which allow a user to verify a mixing result genuinely came from an enclave.