Include the chain's genesis hash in the Glove attestation (#1)
This pins the Glove proofs to the chain they are submitted to and
prevents a dishonest host from manipulating valid Glove proofs by
putting them on other Substrate chains. This is also in-line with the
fact the Glove service only operates on one chain.

Also added tests for verifying the attestation and proofs using samples
taken from on-chain.