Reproducible builds (#3)
Introduced a `glove-build-env` Docker image from which Glove is built.
The image is hosted
[here](https://github.com/projectglove/glove-monorepo/pkgs/container/glove-monorepo%2Fglove-build-env)
in the GitHub org. The build script, `build.sh`, uses a fixed digest of
the image to create the container to ensure the exact same build
environment with the same tool versions are used. This enables different
machines to produce the same enclave image measurement, `PCR0`.

Glove clients can compare the enclave image measurement they get in the
Glove proofs with the value they get by compiling the enclave
themselves, and thus confirm for themselves that a genuine Glove enclave
running on AWS Nitro hardware produced that proof.