Updated README with Glove service REST API doc (#14)
This includes a Polkadot JS sample which creates a signed vote request
payload for the `/vote` endpoint. This endpoint was also updated to
return a JSON object if the request was bad with error information.

`/remove-vote` had to be fixed to be a signed request, otherwise anyone
can remove a vote.

Added `network_name` field to `/info` output, and `network_url` was
renamed to `node_endpoint`

Also added some documentation to the public Rust structs.