Added `verify-vote` command to client (#18)
It takes in an `account` and `poll_index` as arguments and will check if
there is a matching vote on-chain using `subscan`. If so, it will then
make sure there is a valid Glove proof associated with it.

The command also takes in an optional number of `enclave_measurement`s
which are whitelisted Glove enclave versions, and will make sure the
Glove proof is from an enclave with one of these identities. If an
`enclave_measurement` is not provided then it will still valid the
proof, but emphasis the enclave is only a _possible_ Glove enclave. It
then prints instructions on how to clone the correct version tag of this
repo for code audit.