Devops/fix ssh connection (#21)
The initial issue was that I didn't pass the inventory file to ansible.
To ensure that I won't need constantly create new PR to check ansible I
create a new workflow which work on this branch. I'm going to keep it
in, so next time I can just change the branch name and reuse it.