Storing vote requests in DynamoDB table (#23)
The format of the partition key is `POLL#{poll_index}#` and the sort key
is `VOTE#{account}#`. This allows vote requests to be added (including
replacing existing request) and removed in a single operation, and also
allows new items types to be stored in future if necessary.

The service runs a background thread, checking for expired polls and
Glove violators, and this requires knowing all the polls clients have
asked to be voted on. This currently requires scanning the entire table,
something which is generally discouraged. This shouldn't be an issue
with the amount of data that's expected to be stored, but nonetheless,
has been mitigated against with an in-memory cache.

The Ansible deployment hasn't been updated to use DynamoDB and continues
to use the in-memory store. This will be updated separately.