Glove mixing only once near the end of the poll (#24)
The service will mix votes only once when one of the following
conditions are met:

* Near the end of the poll's decision period, regardless of whether it
is confirming or not.
* Near the end of the confirmation period if confirmation will end
within the decision period.

Explanation of these are in the README. "Near" is defined as 160 blocks,
which equates to 15 mins assuming a block time of 6 seconds.

Vote requests and remove vote requests are rejected once votes have been
mixed. A re-mix will only occur if the service detects one of the Glove
participants voting outside of Glove. All of this is checked (and acted
upon) once every 60 seconds on the background thread.

The previous behaviour of mixing regularly, which is necessary for
development and testing purposes, has been preserved behind a
`--regular-mix` flag. The test environment has this enabled.