Various improvements coming out of stress testing (#29)
Created a simple stress testing tool which can submit a large number of
vote requests across all active polls on the network. It takes a start
and end derivation index, and uses a base mnemonic secret phrase to
create a large number of accounts. The only manual process is assigning
tokens with the faucet. The tool invokes the client via its command line
interface. This required moving all of the client logic from its
`main.rs` into `lib.rs`. The verification logic was moved into its own
file.

The following improvements were made as a result:

* Dealing with rate limiting from `subscan` by retrying after
`Retry-After` seconds (if the header is present)
* Returning `429 Too Many Requests` status code on `/vote` and
`/remove-vote` if the DynamoDB operation errors with
`ProvisionedThroughputExceededException` or `RequestLimitExceeded`. The
client has been updated with retry logic to deal with this.
* Using `subscan` to retrieve extrinsic sender account when determining
the voters for a poll (which is done to check for clients voting outside
of Glove), instead of using the `system` `BlockHash` storage. [Node
endpoint](https://wiki.polkadot.network/docs/maintain-endpoints)
services (or at least some of them) seem to be discarding blocks after a
certain amount of time and can't be relied upon for this.
* This extrinsic sender account information is cached globally to
prevent it being retrieved repeatedly by the background thread (which
runs every 60s)

The `node_endpoint` field from `/info` has also been removed. Some
endpoint services require an API key, which would be part of the URL,
and thus having the endpoint in `/info` would be leaking it. Clients can
pick their own endpoint using the `network_name` field.