Service option to exclude polls from voting which belong to certain tracks
`--exclude-tracks` is a comma separated list of track IDs. Vote requests for such tracks return with `TrackNotAllowed` bad request error.

If a track is excluded and there are already vote requests for the track, then a warning will be issued but the requests will be mixed. This is to honour the `/vote` request returning HTTP OK. Further vote requests will not be allowed.