Service option to exclude polls from being voted on which belong to certain tracks (#33)
`--exclude-tracks` is a comma separated list of track IDs. Vote requests
for such tracks return with a `TrackNotAllowed` bad request error.

If a track is excluded and there are already vote requests for the
track, then a warning is issued but the requests will be mixed. This is
to honour the `/vote` request having previously returned HTTP OK.
However, further vote requests will not be allowed.