Client no longer requires node endpoint and subxt reconnect fix (#41)
Previously the client was incorrectly assuming a node endpoint of the
form `wss://${network_name}-rpc.polkadot.io`. In particular, this
doesn't hold for Polkadot and Kusama. The alternative of adding a node
endpoint flag is not workable, as it would make the client harder to
use, especially when it comes to verifying votes. Subscan is being used
instead, which does have a consistent URL format.

The service has also been improved to deal with `RestartNeeded` errors
coming from the RPC client in subxt. This was causing `/poll-info` and
`/vote` to return `500 Internal Server Error`. The subxt RPC client will
now try to automatically reconnect, and if that fails, the end-point
will return `503 Service Unavailable` to indicate a temporary issue.