Test the node RPC endpoint in `/info` (#43)
`/info` gets the current block number from the node endpoint. This
should keep the RPC connection alive (in case the subxt reconnect
functionality isn't working) since `/info` is polled frequently as part
of the health check. And if the subxt call errors then the health check
will cause the service to restart.